视频解码。由于自己手机坏了 ，没法调试。 现在才调试。

+++ Record-master/宠仔圈/RecordView/Decode(解码)/DecodeView.m
@@ -297,9 +297,9 @@ void playCallBack(SystemSoundID ID, void *clientData) {
                 if(decodeStatus == kVTInvalidSessionErr) {
                     NSLog(@"IOS8VT: Invalid session, reset decoder session");
                 } else if(decodeStatus == kVTVideoDecoderBadDataErr) {
-                    NSLog(@"IOS8VT: decode failed status=%d(Bad data)", decodeStatus);
+                    NSLog(@"IOS8VT: decode failed status=%d(Bad data)", (int)decodeStatus);
                 } else if(decodeStatus != noErr) {
-                    NSLog(@"IOS8VT: decode failed status=%d", decodeStatus);
+                    NSLog(@"IOS8VT: decode failed status=%d", (int)decodeStatus);
                 }
                 CFRelease(sampleBuffer);
             }
